pyproject.toml
[project]
name = "py-gsn-relayer"
version = "0.1.0"
description = "Python GSN Relayer Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "web3>=6.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "eth-account>=0.9.0",
    "eth-abi>=4.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
